feat: implement workflow step builders for tenant secret and admin management
