+++ packages/routing/src/admin-resource-routes.ts
@@ -1,13 +1,7 @@
 import type { AdminResourceDefinition } from '@sva/plugin-sdk';
 import { createRoute, redirect, type RootRoute } from '@tanstack/react-router';
 
-import {
-  LEGACY_CONTENT_ALIAS_PREFIX,
-  normalizeLegacyContentHref,
-  readBeforeLoadHref,
-  resolveCanonicalContentAdminRoutePath,
-  withCoreContentAdminResource,
-} from './admin-resource-route-aliases.js';
+import { withCoreContentAdminResource } from './admin-resource-route-aliases.js';
 import {
   adminDetailParamNameByBinding,
   getAdminDetailRoutePath,
@@ -19,6 +13,8 @@ import { createAccountUiRouteGuard, type AccountUiRouteGuardKey } from './accoun
 import { normalizeAdminResourceListSearch } from './admin-resource-search-params.js';
 import type { AppRouteBindings, AppRouteFactory } from './app.routes.shared.js';
 import type { RoutingDiagnosticsHook } from './diagnostics.js';
+
+export { createLegacyContentAliasFactories } from './admin-resource-route-legacy-alias-factories.js';
 
 type BindingKey = keyof AppRouteBindings;
 type UiRouteDefinition = {
@@ -210,28 +206,6 @@ const createAdminResourceRouteDefinitions = (
     ] as const;
   });
 
-const normalizeAliasHref = (href: string, sourcePrefix: string, targetPrefix: string): string => {
-  if (sourcePrefix === LEGACY_CONTENT_ALIAS_PREFIX) {
-    return normalizeLegacyContentHref(href, targetPrefix);
-  }
-
-  if (href === sourcePrefix || href.startsWith(`${sourcePrefix}?`)) {
-    return href.replace(sourcePrefix, targetPrefix);
-  }
-
-  const createSourcePrefix = `${sourcePrefix}/new`;
-  const createTargetPrefix = `${targetPrefix}/new`;
-  if (href === createSourcePrefix || href.startsWith(`${createSourcePrefix}?`)) {
-    return href.replace(createSourcePrefix, createTargetPrefix);
-  }
-
-  if (href.startsWith(`${sourcePrefix}/`)) {
-    return href.replace(`${sourcePrefix}/`, `${targetPrefix}/`);
-  }
-
-  return targetPrefix;
-};
-
 export const createAdminResourceRouteFactories = (
   bindings: AppRouteBindings,
   resources: readonly AdminResourceDefinition[],
@@ -254,36 +228,3 @@ export const createAdminResourceRouteFactories = (
         });
       }
   );
-
-export const createLegacyContentAliasFactories = (
-  resources: readonly AdminResourceDefinition[] = []
-): readonly AppRouteFactory[] => {
-  const canonicalContentPath = resolveCanonicalContentAdminRoutePath(resources);
-  const aliasMappings = [
-    {
-      sourcePrefix: LEGACY_CONTENT_ALIAS_PREFIX,
-      targetPrefix: canonicalContentPath,
-    },
-    ...resources
-      .filter((resource) => resource.guard === 'content')
-      .map((resource) => ({
-        sourcePrefix: `/plugins/${resource.basePath}`,
-        targetPrefix: toAdminRoutePath(resource.basePath),
-      })),
-  ] as const;
-
-  return aliasMappings.flatMap(({ sourcePrefix, targetPrefix }) =>
-    [sourcePrefix, `${sourcePrefix}/new`, `${sourcePrefix}/$contentId`].map((path) => (rootRoute: RootRoute) =>
-      createRoute({
-        getParentRoute: () => rootRoute,
-        path,
-        beforeLoad: (options) => {
-          const href = readBeforeLoadHref(options);
-          const normalizedHref = normalizeAliasHref(href, sourcePrefix, targetPrefix);
-          throw redirect({ href: normalizedHref });
-        },
-        component: () => null,
-      })
-    )
-  );
-};
